feat: add ProductSeeDistributionSotcksLocation component
