+++ packages/apps/stock/src/components/organisms/index.js
@@ -56,3 +56,4 @@ export {default as SupplierArrivalLineDetailFixedItems} from './SupplierArrivalL
 export {default as SupplierArrivalLineDetailQuantityCard} from './SupplierArrivalLineDetailQuantityCard/SupplierArrivalLineDetailQuantityCard';
 export {default as SupplierProductInfo} from './SupplierProductInfo/SupplierProductInfo';
 export {default as SupplierArrivalAddTracking} from './SupplierArrivalAddTracking/SupplierArrivalAddTracking';
+export {default as ProductSeeDistributionStocksLocation} from './ProductSeeDistributionStocksLocation/ProductSeeDistributionStocksLocation';
